Bypass login activity with valid jwt
